+++ modules/display/geotk-wizards-swing/src/main/java/org/geotoolkit/resources/Wizards.java
@@ -50,19 +50,19 @@ public final class Wizards extends IndexedResourceBundle {
         /**
          * Calculation in progress...
          */
-        public static final int CALCULATION_PROGESSING = 0;
+        public static final short CALCULATION_PROGESSING = 0;
 
         /**
          * <html><blockquote>The database has been created. Check the connection parameters in the
          * preferences menu. Make sure that the data file directory and the time zone are set to the
          * proper values.</blockquote></html>
          */
-        public static final int COVERAGE_DATABASE_CREATED = 1;
+        public static final short COVERAGE_DATABASE_CREATED = 1;
 
         /**
          * Creates the tables for the "coverages" schema in an initially empty database.
          */
-        public static final int COVERAGE_DATABASE_DESC = 2;
+        public static final short COVERAGE_DATABASE_DESC = 2;
 
         /**
          * <html><b>Notes:</b><ul><li>Roles shall be created only once for a given server. If a
@@ -71,78 +71,78 @@ public final class Wizards extends IndexedResourceBundle {
          * href="http://www.epsg.org">http://www.epsg.org</a>) is sufficient and can be shared by
          * different Coverages database. However having multiple copies is okay.</li></ul></html>
          */
-        public static final int COVERAGE_DATABASE_NOTES_1 = 3;
+        public static final short COVERAGE_DATABASE_NOTES_1 = 3;
 
         /**
          * New coverage database
          */
-        public static final int COVERAGE_DATABASE_TITLE = 4;
+        public static final short COVERAGE_DATABASE_TITLE = 4;
 
         /**
          * Copy the EPSG database
          */
-        public static final int CREATE_EPSG = 5;
+        public static final short CREATE_EPSG = 5;
 
         /**
          * Create the “{0}” and “{1}” roles
          */
-        public static final int CREATE_ROLES_2 = 6;
+        public static final short CREATE_ROLES_2 = 6;
 
         /**
          * Creating the mosaic
          */
-        public static final int CREATING_MOSAIC = 7;
+        public static final short CREATING_MOSAIC = 7;
 
         /**
          * Creating the {0} schema.
          */
-        public static final int CREATING_SCHEMA_1 = 8;
+        public static final short CREATING_SCHEMA_1 = 8;
 
         /**
          * A server and a database must be specified.
          */
-        public static final int DATABASE_REQUIRED = 9;
+        public static final short DATABASE_REQUIRED = 9;
 
         /**
          * Define pyramid tiling
          */
-        public static final int DEFINE_PYRAMID_TILING = 10;
+        public static final short DEFINE_PYRAMID_TILING = 10;
 
         /**
          * The wizard has now enough informations for creating the mosaic. Press "Finish" to confirm.
          */
-        public static final int ENOUGH_INFORMATION = 11;
+        public static final short ENOUGH_INFORMATION = 11;
 
         /**
          * Geotoolkit.org web site
          */
-        public static final int GEOTK_SITE = 12;
+        public static final short GEOTK_SITE = 12;
 
         /**
          * Geotoolkit.org wizards
          */
-        public static final int GEOTK_WIZARDS = 13;
+        public static final short GEOTK_WIZARDS = 13;
 
         /**
          * The selected tiles can not make a single mosaic.
          */
-        public static final int INVALID_MOSAIC_LAYOUT = 14;
+        public static final short INVALID_MOSAIC_LAYOUT = 14;
 
         /**
          * Read a potentially big image (which may be splitted in many tiles at the same resolution)
          * and write a set of smaller tiles of given size and using different subsamplings.
          */
-        public static final int MOSAIC_DESC = 15;
+        public static final short MOSAIC_DESC = 15;
 
         /**
          * Mosaic generator
          */
-        public static final int MOSAIC_TITLE = 16;
+        public static final short MOSAIC_TITLE = 16;
 
         /**
          * At least one tile must be selected.
          */
-        public static final int NO_SELECTED_TILES = 17;
+        public static final short NO_SELECTED_TILES = 17;
 
         /**
          * <html>Specify the directory which contain the <code>postgis.sql</code> and
@@ -150,44 +150,44 @@ public final class Wizards extends IndexedResourceBundle {
          * server, make sure that the files specified below are identical to the files on the
          * server.</html>
          */
-        public static final int POSTGIS_DIRECTORY = 18;
+        public static final short POSTGIS_DIRECTORY = 18;
 
         /**
          * Remove opaque border
          */
-        public static final int REMOVE_OPAQUE_BORDER = 19;
+        public static final short REMOVE_OPAQUE_BORDER = 19;
 
         /**
          * Select source tiles
          */
-        public static final int SELECT_SOURCE_TILES = 20;
+        public static final short SELECT_SOURCE_TILES = 20;
 
         /**
          * Select directories and install the NADCON and EPSG data. This setup is optional. If
          * executed, the setting will be remembered for all subsequent Geotoolkit.org usage.
          */
-        public static final int SETUP_DESC = 21;
+        public static final short SETUP_DESC = 21;
 
         /**
          * Geotoolkit.org Setup
          */
-        public static final int SETUP_TITLE = 22;
+        public static final short SETUP_TITLE = 22;
 
         /**
          * Set as the default {0} database
          */
-        public static final int SET_AS_DEFAULT_1 = 23;
+        public static final short SET_AS_DEFAULT_1 = 23;
 
         /**
          * <html>Use the <cite>Preferences</cite> menu for specifying an existing database,<br>or use
          * the <cite>New coverage database</cite> menu for creating a new database.</html>
          */
-        public static final int UNSPECIFIED_COVERAGES_DATABASE = 24;
+        public static final short UNSPECIFIED_COVERAGES_DATABASE = 24;
 
         /**
          * Writing the mosaic
          */
-        public static final int WRITING_MOSAIC = 25;
+        public static final short WRITING_MOSAIC = 25;
     }
 
     /**
@@ -217,7 +217,7 @@ public final class Wizards extends IndexedResourceBundle {
      * @return The string for the given key.
      * @throws MissingResourceException If no object for the given key can be found.
      */
-    public static String format(final int key) throws MissingResourceException {
+    public static String format(final short key) throws MissingResourceException {
         return getResources(null).getString(key);
     }
 }
